Moving operator code to a designated folder so that webhook work can also be included in this repo
