+++ operator/deploy/crds/onepassword.com_onepassworditems_crd.yaml
