+++ operator/deploy/crds/onepassword.com_v1_onepassworditem_cr.yaml
