+++ operator/deploy/operator_multi_namespace_example.yaml
