+++ operator/deploy/permissions_multi_namespace_example.yaml
