+++ operator/pkg/apis/onepassword/v1/onepasswordsecret_types.go
