+++ operator/pkg/apis/onepassword/v1/register.go
