+++ operator/pkg/onepassword/connect_setup_test.go
