+++ operator/pkg/onepassword/items.go
