+++ operator/version/version.go
